+++ NEWS.md
@@ -6,6 +6,8 @@
 
 ## Minor improvements and bug fixes
 
+* `$new()` and `plm_template()` gain a new parameter `credit_roles` to facilitate the use of the [Contributor Roles Taxonomy](https://credit.niso.org). 
+
 * `PlumeQuarto` now handles roles via the `roles` YAML key (#5).
 
 * New method `get_orcids()` that returns authors' ORCID.